requirements.txt
fastapi
uvicorn
requests
pandas
numpy
traffic
python-dotenv